feat: move the creation of connections_cell from ibc-test to relayer core
