+++ tools/ibc-test/src/consts.rs
@@ -5,15 +5,15 @@ pub const SUDT_CODE_HASH: H256 =
 pub const CONNECTION_CODE_HASH: H256 =
     h256!("0xe802fa98adeed078eb39225f6d7759ceeeecc32fa28942f2d2c6a158ec40cebb");
 pub const CHANNEL_CODE_HASH: H256 =
-    h256!("0x0994bee0f1e606c7461c80b9b768c3163679203ed342f3d55890cb9c025251a0");
+    h256!("0x2b665478159b4759370dc98b4f327767de23f9e675cd546236be99daad8f0f32");
 
 pub const SUDT_TYPE_ARGS: H256 =
     h256!("0xf49ce32397c6741998b04d7548c5ed372007424daf67ee5bfadaefec3c865781");
 pub const CONNECTION_TYPE_ARGS: H256 =
     h256!("0x9d46ad77b8f5107d5898d7e7ad82be206c4fd95f562004efe32488605e8846c7");
 pub const CHANNEL_TYPE_ARGS: H256 =
-    h256!("0xfa8b909ec087b37c60ed78b5406f0980a5de0ebf3cee1d0e6c03699ae81fc09d");
+    h256!("0xf317187ec28d6a62b9e4642f03fb1196b15fcce5b70c552f36b59adeea28d44a");
 pub const PACKET_TYPE_ARGS: H256 =
-    h256!("0x4048ccfa831feae246c6652c0414c65e7696d4981362f5b94f94d40af86b403d");
+    h256!("0x69eda031210bb4eb12dd7c18cd01e297a1d36f035312fde1aa698da302e16441");
 pub const CLIENT_TYPE_ARGS: H256 =
-    h256!("0x99c77b7a7e35e400bc78e6dc88191c0ab49dfd945134d99952155d08a2f6d96b");
+    h256!("0x30ca27ae4db0158f0ab491896fac626ffc7aab24f364fff20d74b189814435f7");
