--- tools/ibc-test/txs/create_connection_sudt.json
@@ -1,1 +1,0 @@
-{"id": 5, "jsonrpc": "2.0", "method": "send_transaction", "params": [{"version":"0x0","cell_deps":[{"out_point":{"tx_hash":"0xb86b4284274692f8f3e67fc0001386477f0c4b6aed104c7abb98a6cf7636e47b","index":"0x1"},"dep_type":"code"},{"out_point":{"tx_hash":"0xb86b4284274692f8f3e67fc0001386477f0c4b6aed104c7abb98a6cf7636e47b","index":"0x0"},"dep_type":"code"},{"out_point":{"tx_hash":"0x29ed5663501cd171513155f8939ad2c9ffeb92aa4879d39cde987f8eb6274407","index":"0x0"},"dep_type":"dep_group"},{"out_point":{"tx_hash":"0xd9812f0aaa786955967cf9ad69b959d97a9ea5a7b3df76b6f900c6ad622a2ad5","index":"0x0"},"dep_type":"code"}],"header_deps":[],"inputs":[{"since":"0x0","previous_output":{"tx_hash":"0xb86b4284274692f8f3e67fc0001386477f0c4b6aed104c7abb98a6cf7636e47b","index":"0x2"}}],"outputs":[{"capacity":"0x271d94900","lock":{"code_hash":"0xe802fa98adeed078eb39225f6d7759ceeeecc32fa28942f2d2c6a158ec40cebb","hash_type":"type","args":"0xdea4c5682c1acf8421dff9b51589311137e750e8d7e3588095e59dd47a34cd7d"},"type":null},{"capacity":"0x34e62ce00","lock":{"code_hash":"0x9bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8","hash_type":"type","args":"0x470dcdc5e44064909650113a274b3b36aecb6dc7"},"type":{"code_hash":"0xcf6e0c0148123081af1deda0ef162d39cfdfe1ea6565d3689009c1f3562a5e82","hash_type":"type","args":"0xc219351b150b900e50a7039f1e448b844110927e5fd9bd30425806cb8ddff1fd"}},{"capacity":"0x16bcc41e90000","lock":{"code_hash":"0x9bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8","hash_type":"type","args":"0x470dcdc5e44064909650113a274b3b36aecb6dc7"},"type":null}],"outputs_data":["0x83795d737ad7359b0f3f87eba317e87c7eacba1ea3e21a7f00764c91ad265d2d","0x10270000000000000000000000000000","0x"],"witnesses":["0x5d00000010000000550000005500000041000000ec00cb385cf19e62cc71945147314ea71eec7137e2d31720f032ac21591bf0002af0b24eb154cb1e43b16f20c3c2f5683e4a38a47e0ae1a1505410863b93936f0104000000c38080c0","0x10000000100000001000000010000000","0x1900000010000000100000001000000005000000c401c281c0"]}]}
