+++ tools/ibc-test/txs/issue_sudt.json
@@ -1,0 +1,1 @@
+{"id": 5, "jsonrpc": "2.0", "method": "send_transaction", "params": [{"version":"0x0","cell_deps":[{"out_point":{"tx_hash":"0x29ed5663501cd171513155f8939ad2c9ffeb92aa4879d39cde987f8eb6274407","index":"0x0"},"dep_type":"dep_group"},{"out_point":{"tx_hash":"0xd9812f0aaa786955967cf9ad69b959d97a9ea5a7b3df76b6f900c6ad622a2ad5","index":"0x0"},"dep_type":"code"}],"header_deps":[],"inputs":[{"since":"0x0","previous_output":{"tx_hash":"0x40623599ed5017c3a2f959a4521ba96097451491d97333f8eb41e597ebc4791c","index":"0x2"}}],"outputs":[{"capacity":"0x34e62ce00","lock":{"code_hash":"0x9bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8","hash_type":"type","args":"0x470dcdc5e44064909650113a274b3b36aecb6dc7"},"type":{"code_hash":"0xcf6e0c0148123081af1deda0ef162d39cfdfe1ea6565d3689009c1f3562a5e82","hash_type":"type","args":"0xc219351b150b900e50a7039f1e448b844110927e5fd9bd30425806cb8ddff1fd"}},{"capacity":"0x16bcc41e90000","lock":{"code_hash":"0x9bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8","hash_type":"type","args":"0x470dcdc5e44064909650113a274b3b36aecb6dc7"},"type":null}],"outputs_data":["0x10270000000000000000000000000000","0x"],"witnesses":["0x550000001000000055000000550000004100000099ee0f499df23727a4075b470cef9071c2bb4fad2f3b443d10d4a02a0f4bd15b270343847028e8b92e18d3dae9ca6bf17ebcbffffa62a3a6b6bbdc30a309892301"]}]}
